pyproject.toml
[project]
name = "agenthalo"
version = "0.1.0"
description = ""
authors = [
    {name = "Dennis Won", email = "dennis.won@magic.link"},
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "annotated-types (>=0.7.0,<0.8.0)",
    "anyio (>=4.8.0,<5.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "dstack-sdk (>=0.1.5,<0.2.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "exceptiongroup (>=1.2.2,<2.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "fastapi-cli (>=0.0.7,<0.0.8)",
    "h11 (>=0.14.0,<0.15.0)",
    "httpcore (>=1.0.7,<2.0.0)",
    "httptools (>=0.6.4,<0.7.0)",
    "idna (>=3.10,<4.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "mdurl (>=0.1.2,<0.2.0)",
    "pydantic-core (>=2.31.1,<3.0.0)",
    "pygments (>=2.19.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "rich (>=13.9.4,<14.0.0)",
    "rich-toolkit (>=0.13.2,<0.14.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "starlette (>=0.46.1,<0.47.0)",
    "typer (>=0.15.2,<0.16.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "uvloop (>=0.21.0,<0.22.0)",
    "watchfiles (>=1.0.4,<2.0.0)",
    "base58 (>=2.1.1,<3.0.0)",
]

[tool.poetry.group.ai.dependencies]
transformers = "4.47.1"  # fixed smolagents dependency
litellm = "1.55.10"  # newer versions via smolagents will not work on windows
instructor = "1.7.2"
smolagents = "1.2.2"

[tool.poetry.group.clients.dependencies]
python-telegram-bot = "21.10"

[tool.poetry.group.web3.dependencies]
web3 = "6.14.0"
solana = "0.36.0"
web3-ethereum-defi = "0.28"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20241230"
ruff = "^0.9.5"
mypy = "^1.15.0"
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.4"
types-requests = ">=2.31.0.6"
python-dotenv = "^1.0.1"
eth-typing = "3.5.2"
pytest-html = "^4.1.1"
pytest-cov = "^6.0.0"

[tool.poetry]
include = [
    { path = "config/default.yaml", format = "wheel" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

#[tool.pytest]
#pythonpath = "."

[tool.pytest.ini_options]
log_cli = "True"
log_cli_level = "DEBUG"
log_cli_format = "[%(asctime)s %(levelname)s %(threadName)s %(name)s:%(funcName)s:%(lineno)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S%z"

[tool.black]
line-length = 120

[tool.mypy]
exclude = "(?x)(venv|docs|tmp)"

explicit_package_bases = true
no_implicit_optional = true
disallow_untyped_defs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["smolagents.*", "eth_defi.*", "dstack_sdk.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
exclude = ["__init__.py", "docs", "tmp", "venv"]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
py_version = 311
src_paths = ["src"]
skip = ["__init__.py"]
skip_glob = ["**/.venv/*", "**/venv/*", "**/docs/*", "**/stubs/*", "**/tests/*"]
